+++ src/main/java/nextstep/ladder/domain/Name.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.Objects;
 import nextstep.ladder.exception.OverNameLengthLimitException;
